refactor: rearrange packages and classes by feature
This greatly improves discoverability and overview of the application. Currently the upside is not as big, but when adding more and more features the whole project structure will be easier to overview.
