+++ lib/firebase/services/firebase_service.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:get_it/get_it.dart';
-import 'package:lifting_progress_tracker/firebase_options.dart';
-import 'package:lifting_progress_tracker/services/firebase/firestore_service.dart';
+import 'package:lifting_progress_tracker/firebase/firebase_options.dart';
+import 'package:lifting_progress_tracker/firebase/services/firestore_service.dart';
 import 'package:logging/logging.dart';
 
 class FirebaseService {
