+++ lib/firebase/services/firestore_service.dart
@@ -1,5 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:lifting_progress_tracker/types/types.dart';
+import 'package:lifting_progress_tracker/firebase/types.dart';
 import 'package:logging/logging.dart';
 
 class FirestoreService {
