+++ lib/pages/calendar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:lifting_progress_tracker/constants/localization.dart';
-import 'package:lifting_progress_tracker/widgets/navigation_button.dart';
+import 'package:lifting_progress_tracker/core/constants/localization.dart';
+import 'package:lifting_progress_tracker/core/widgets/navigation_button.dart';
 
 class CalendarPage extends StatelessWidget {
   @override
