+++ lib/training_plan/models/plan_entry.dart
