+++ lib/training_plan/widgets/table/add_entry_button.dart
