+++ lib/training_plan/widgets/table/entry_removal_button.dart
