+++ lib/training_plan/widgets/table/workout_table.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:lifting_progress_tracker/constants/localization.dart';
-import 'package:lifting_progress_tracker/models/plan_entry.dart';
-import 'package:lifting_progress_tracker/providers/table_provider.dart';
-import 'package:lifting_progress_tracker/widgets/table/add_entry_button.dart';
-import 'package:lifting_progress_tracker/widgets/table/entry_removal_button.dart';
-import 'package:lifting_progress_tracker/widgets/table/table_text_field.dart';
+import 'package:lifting_progress_tracker/core/constants/localization.dart';
+import 'package:lifting_progress_tracker/training_plan/models/plan_entry.dart';
+import 'package:lifting_progress_tracker/training_plan/provider/table_provider.dart';
+import 'package:lifting_progress_tracker/training_plan/widgets/table/add_entry_button.dart';
+import 'package:lifting_progress_tracker/training_plan/widgets/table/entry_removal_button.dart';
+import 'package:lifting_progress_tracker/training_plan/widgets/table/table_text_field.dart';
 import 'package:provider/provider.dart';
 
 const _headerStyle = TextStyle(fontWeight: FontWeight.bold);
